refactor: optimized the code
