+++ src/constants/data/data.ts
@@ -1,0 +1,15 @@
+import { Product } from 'constants/types/types';
+
+export const sortValues: Array<string> = [
+    'default',
+    'price ASC',
+    'price DESC',
+    'rating ASC',
+    'rating DESC',
+    'dicount ASC',
+    'dicount DESC',
+];
+
+export const gridValues: Array<string> = ['3', '4'];
+
+export const cartArray: Array<Product> = [];
