+++ src/pages/main/catalog/catalog.ts
@@ -1,5 +1,6 @@
 import * as utils from '../../../utils/index';
 import './catalog.scss';
+import { sortValues, gridValues } from '../../../constants/data/data';
 
 export function createCatalog(): HTMLElement {
     const catalog = utils.createElement('section', 'catalog');
@@ -36,25 +37,7 @@ function createSortSelection(): HTMLElement {
     const sortSelection = utils.createElement('select', 'catalog__sort-selection', 'selection');
     const sortDescription = utils.createElement('span', 'catalog__sort-description');
 
-    //можно вынести в константы
-    const sortValues: Array<string> = [
-        'default',
-        'price ASC',
-        'price DESC',
-        'rating ASC',
-        'rating DESC',
-        'dicount ASC',
-        'dicount DESC',
-    ];
-
-    sortValues.forEach((item) => {
-        const option = document.createElement('option');
-
-        option.setAttribute('value', item);
-        option.textContent = item;
-
-        sortSelection.append(option);
-    });
+    utils.createSelectOptions(sortValues, sortSelection);
 
     sortDescription.textContent = 'Sort by ';
 
@@ -72,6 +55,7 @@ function createNumberOfProducts(): HTMLElement {
     span.textContent = '0';
 
     p.append(span);
+
     return p;
 }
 
@@ -90,17 +74,7 @@ function createSelectGrid(): HTMLElement {
     const gridDescription = utils.createElement('span', 'catalog__grid-Description');
     const gridSelection = utils.createElement('select', 'catalog__grid-selection', 'selection');
 
-    //можно вынести в константы
-    const gridValues: Array<string> = ['3', '4'];
-
-    gridValues.forEach((item) => {
-        const option = document.createElement('option');
-
-        option.setAttribute('value', item);
-        option.textContent = item;
-
-        gridSelection.append(option);
-    });
+    utils.createSelectOptions(gridValues, gridSelection);
 
     gridDescription.textContent = 'Show by ';
